requirements.txt
streamlit==1.50.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.1
pydantic>=2.0.0
litellm>=1.0.0
supabase>=2.28.0
pywebview>=5.0.0
paramiko>=3.0.0
python-dotenv>=1.0.0
requests>=2.30.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-multipart>=0.0.6
protobuf<5.0.0
apscheduler>=3.10.0
reportlab>=4.0.0
pynput>=1.7.6
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
slack_sdk>=3.0.0
pyautogui>=0.9.54
Pillow>=10.0.0
openai>=1.0.0
duckduckgo-search>=6.0.0
pypdf>=4.0.0
opencv-python>=4.8.0
